+++ packages/gov-ui-kit/src/modules/components/asset/assetDataListItem/assetDataListItemStructure/assetDataListItemStructure.tsx
@@ -1,7 +1,6 @@
 import classNames from 'classnames';
 import type React from 'react';
-import { useMemo } from 'react';
-import { Avatar, DataList, NumberFormat, Tag, formatterUtils, type IDataListItemProps } from '../../../../../core';
+import { Avatar, DataList, NumberFormat, formatterUtils, type IDataListItemProps } from '../../../../../core';
 import { useGukModulesContext } from '../../../gukModulesProvider';
 
 export type IAssetDataListItemStructureProps = IDataListItemProps & {
@@ -25,38 +24,14 @@ export type IAssetDataListItemStructureProps = IDataListItemProps & {
      * The fiat price of the asset.
      */
     fiatPrice?: number | string;
-    /**
-     * The price change in percentage of the asset.
-     * @default 0
-     */
-    priceChange?: number;
 };
 
 export const AssetDataListItemStructure: React.FC<IAssetDataListItemStructureProps> = (props) => {
-    const { logoSrc, name, amount, symbol, fiatPrice, priceChange = 0, className, ...otherProps } = props;
+    const { logoSrc, name, amount, symbol, fiatPrice, className, ...otherProps } = props;
 
     const { copy } = useGukModulesContext();
 
-    const fiatAmount = Number(amount) * Number(fiatPrice ?? 0);
-
-    const fiatAmountChanged = useMemo(() => {
-        if (!fiatPrice || !priceChange) {
-            return 0;
-        }
-
-        const oldFiatAmount = (100 / (priceChange + 100)) * fiatAmount;
-
-        return fiatAmount - oldFiatAmount;
-    }, [fiatAmount, fiatPrice, priceChange]);
-
-    const changedAmountClasses = classNames(
-        'text-sm font-normal leading-tight md:text-base',
-        { 'text-success-800': fiatAmountChanged > 0 },
-        { 'text-neutral-500': fiatAmountChanged === 0 },
-        { 'text-critical-800': fiatAmountChanged < 0 },
-    );
-
-    const tagVariant = priceChange > 0 ? 'success' : priceChange < 0 ? 'critical' : 'neutral';
+    const fiatAmount = Number(amount) * Number(fiatPrice);
 
     const formattedAmount = formatterUtils.formatNumber(amount, {
         format: NumberFormat.TOKEN_AMOUNT_SHORT,
@@ -65,46 +40,25 @@ export const AssetDataListItemStructure: React.FC<IAssetDataListItemStructurePro
 
     const formattedPrice = formatterUtils.formatNumber(fiatAmount, {
         format: NumberFormat.FIAT_TOTAL_SHORT,
-        fallback: '-',
-    });
-
-    const formattedPriceChanged = formatterUtils.formatNumber(fiatAmountChanged, {
-        format: NumberFormat.FIAT_TOTAL_SHORT,
-        withSign: true,
-    });
-
-    const formattedPriceChangedPercentage = formatterUtils.formatNumber(priceChange / 100, {
-        format: NumberFormat.PERCENTAGE_LONG,
-        withSign: true,
+        fallback: copy.assetDataListItemStructure.unknown,
     });
 
     return (
-        <DataList.Item className={classNames('flex items-center gap-x-3 py-3 md:py-5', className)} {...otherProps}>
-            <Avatar src={logoSrc} responsiveSize={{ md: 'md', sm: 'sm' }} className="block" />
-            <div className="flex w-full min-w-0 shrink items-center justify-between gap-3">
-                <div className="flex flex-col gap-y-1 truncate">
-                    <span className="truncate text-base leading-tight text-neutral-800 md:text-lg">{name}</span>
-                    <p className="truncate text-sm leading-tight text-neutral-500 md:text-base">
-                        <span>{formattedAmount} </span>
-                        <span className="truncate">{symbol}</span>
-                    </p>
-                </div>
-                <div className="flex flex-col items-end justify-center gap-y-1">
-                    {fiatPrice ? (
-                        <>
-                            <span className="text-base leading-tight text-neutral-800 md:text-lg">
-                                {formattedPrice}
-                            </span>
-                            <div className="flex items-center gap-x-1">
-                                <span className={changedAmountClasses}>{formattedPriceChanged}</span>
-                                <Tag label={formattedPriceChangedPercentage!} variant={tagVariant} />
-                            </div>
-                        </>
-                    ) : (
-                        <span className="text-sm leading-tight text-neutral-800 md:text-base">
-                            {copy.assetDataListItemStructure.unknown}
-                        </span>
-                    )}
+        <DataList.Item
+            className={classNames('flex items-center justify-between gap-x-3 py-3 md:py-5', className)}
+            {...otherProps}
+        >
+            <div className="flex min-w-0 items-center gap-3">
+                <Avatar src={logoSrc} responsiveSize={{ md: 'md', sm: 'sm' }} className="block shrink-0" />
+                <span className="truncate text-base leading-tight text-neutral-800 md:text-lg">{name}</span>
+            </div>
+            <div className="flex min-w-0 gap-x-2 text-right">
+                <div className="flex min-w-0 flex-col gap-y-1">
+                    <span className="text-base leading-tight text-neutral-800 md:text-lg">{formattedPrice}</span>
+                    <div className="flex items-center gap-1">
+                        <p className="text-sm leading-tight text-neutral-500 md:text-base">{formattedAmount}</p>
+                        <p className="truncate text-sm leading-tight text-neutral-500 md:text-base">{symbol}</p>
+                    </div>
                 </div>
             </div>
         </DataList.Item>