+++ packages/gov-ui-kit/src/core/components/illustrations/illustrationObject/illustrationObject.stories.tsx
@@ -1,10 +1,12 @@
 import type { Meta, StoryObj } from '@storybook/react';
+import style from '../index.css?raw';
 import { IllustrationObject } from './illustrationObject';
 
 const meta: Meta<typeof IllustrationObject> = {
     title: 'Core/Components/Illustrations/IllustrationObject',
     component: IllustrationObject,
     parameters: {
+        style,
         design: {
             type: 'figma',
             url: 'https://www.figma.com/file/jfKRr1V9evJUp1uBeyP3Zz/v1.0.0?type=design&node-id=8874-14443&mode=dev',