+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Update Storybook branding and repository readme file
+
 ## [1.0.59] - 2024-12-11
 
 ### Fixed