+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Update Storybook branding and repository readme file
+- Bump `softprops/action-gh-release` from 2.1.0 to 2.2.0
 
 ## [1.0.59] - 2024-12-11