+++ packages/gov-ui-kit/CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Bump `softprops/action-gh-release` from 2.1.0 to 2.2.0
 - Bump `@rollup/plugin-node-resolve` from 15.3.0 to 16.0.0
 - Update minor and patch NPM dependencies
+- Update `AssetDataListItemStructure` module component to remove the `priceChange` property and related logic for
+  calculating percentage change
 
 ## [1.0.59] - 2024-12-11