+++ packages/gov-ui-kit/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Remove `hash` property on the `AssetTransfer` module component and add optional `assetAddress` property
 - Add link to the block explorer for members on the `ProposalActionChangeMembers` module component
 - Add receiver block explorer link to the `ProposalActionTokenMint` module component
+- Update minor and patch NPM dependencies
 
 ## [1.0.60] - 2024-12-18