+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.61] - 2024-12-29
+
 ### Added
 
 - Implement `StyleBlock` component to document the components-specific customisations